@@ -1,7 +1,7 @@
 ﻿/**
  * @type {import("prettier").Config}
  */
-module.exports = {
+export default {
   arrowParens: "always",
   bracketSpacing: true,
   editorconfig: true,
